fix(AccountsController.java): disable password validator
The password validator does not work correctly
at all. If the password is correct and passes all
validation, then there are no problems. However,
when the incorrect password is supplied the first
time, instead of creating a completely new list
of errors in subsequent requests, the password 
validator adds errors to the errors from the
previous requests. This results in the password
always being 'incorrect' in all subsequent runs
and does not get resolved until the server is
restarted or dev tools restart.